Support coloured output for Windows console
The Windows Console doesn't natively support ANSI escape sequences, but
it does include API functions providing most of the required support.
In order to maintain the option of compiling with MSVC, a custom wrapper
is implemented for the small number of required escape sequences, rather
than using MSYS's ANSI layer.

The need to post-process coloured messages means that the format
signatures of various functions in OpamConsole have changed from printf
to sprintf form and it is imperative that all output to the Console
which includes coloured text goes through one of OpamConsole's functions
(in particular, print_string, Printf.printf, output_string std* must not
be used).

Particular care has been taken to ensure that OpamConsole.log and
OpamConsole.slog behave in the same way; specifically this means that
strings logged should never include ANSI escape sequences because these
cannot be translated.

(note that an earlier attempt at doing this used a pipe to write the
data via kfprintf in order to maintain the same %a behaviour and then read
the string back from the pipe, but this leads to problems if the data being
written is larger than the pipe's buffer and it seems excessive to
introduce multi-threading or two processes solely to handle debugging
console output)

Signed-off-by: David Allsopp <david.allsopp@metastack.com>