Support for tools needing Cygwin paths
Various Cygwin tools (notably tar and rsync) don't recognise Windows
bashslash-style paths and need to have the path converted using the
cygpath utility.

Add two functions to OpamSystem to support this:

1. OpamSystem.is_cygwin_variant uses cygcheck to determine if a given
program is the Cygwin-compiled version (cygcheck shows DLLs linked, so
the presence of cygwin1.dll is taken to imply a Cygwin-based command).
Note that for a Cygwin build of OPAM (and indeed for any other Unix-like
build), this function always returns false. For a native Cygwin build,
the user should specify Cygwin-style paths.
2. OpamSystem.apply_cygpath converts its argument to a Cygwin-style path
using the cygpath utility

Signed-off-by: David Allsopp <david.allsopp@metastack.com>