Add DEVELOPER_CACHE configure variable
Passing DEVELOPER_CACHE automatically enables developer mode and allows
the developer to specify a local directory in which remote files
(urls.txt and index.tar.gz in particular) can be placed.
OpamDownload.really_download checks an index placed in this directory
before invoking wget/curl and simply copies the file if it's found.

OPAM automatically adds new files downloaded to this cache, as long as
the directory exists when OPAM is started.