Support UTF-8 display in Windows console
The Windows Console Host supports UTF-8 as long as a TrueType font has
been selected (two are installed by default). Console logic updated to
detect this font support and enable UTF-8 output as appropriate.

At present (and, based on the Windows 10 previews, we can assume for a
while), the Console does not support Unicode surrogate pairs, so sadly
there will be no Camel symbol for OPAM-on-Windows for the foreseeable
future :'( (additionally, the only common font containing it is Segoe UI
Symbol which is not usable as a Console font).

A bug in the C standard library write function means that OCaml's output
functions (and hence Printf) cannot be used with UTF-8, so output is
switched to use WriteConsole directly.

Although Windows has two suitable fonts (Lucida Console and Consolas)
containing Unicode glyphs, neither of these fonts contain the glyphs
used by OPAM's solver. Unsurprisingly, Windows font mapping for symbols
is also vastly inferior compared with Linux, so instead of requiring the
user to find a suitable console font, OPAM detects the glyphs supported
by the font and falls back on some suitable alternatives if the defaults
are not available.