Windows manipulation of parent process environment.
The Windows Command prompt doesn't have an obvious equivalent to sh's
eval, beyond the hideously ugly generation of a temporary .cmd script
without setlocal at the start.

Windows, however, has some moderately devious mechanisms to use an
alternative. OpamStd.Win32.parent_putenv is the same as Unix.putenv,
except that it manipulates the parent-process's environment (which, for
opam.exe, will be the Command Prompt).

The implementation relies on process injection which and takes into
account Windows-on-Windows (i.e. 64/32-bit issues). In 64-bit Windows, a
injecting 32-bit code into a 64-bit process will silently fail and
injecting 64-bit code into a 32-bit process will kill the process. It is
relatively easy for a 64-bit process to inject 32-bit code into its
parent but it's incredibly hard to do the reverse and both mechanisms
are fairly brittle (the curious reader is invited to search online for
Heaven's Gate).

In order to get around this problem, Windows builds of OPAM compile and
install an additional small utility called opam-putenv. This utility is
a 64-bit executable if OPAM is 32-bit and vice versa. An warning is
displayed if the complementary compiler is not found - in this instance,
OPAM must be run in a shell using the same architecture (meaning on
64-bit Windows, a 32-bit build of OPAM must be run from
%windir%\SysWOW64\cmd.exe or it will display an error message).

This implementation has the benefit, for Windows users, that unlike in
other shells, it's not necessary to run opam config env at all after
opam init, opam switch, etc. because OPAM can configure the environment
instantly. This may prove to be the only place where Windows/CMD has the
upper hand...

This technique was inspired by Bill Stewart's editvar
(see http://www.westmesatech.com/editv.html), a handy little utility
primarily designed for prompting for passwords in Windows CMD scripts,
which outputs the password entered by manipulating the parent's
environment.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>