Workaround conflict over link command
The Microsoft Linker is called link which has the potential to conflict
with coreutils link, depending on the order of directories in PATH. By
default, coreutils link will win, and compilation is broken on MSVC
OCaml.

Simply altering /etc/profile to put the Cygwin directories later isn't
an option as commands like find and sort then get overridden by
Microsoft's "alternative" implementations in system32.

Introduce detection in configure so that if multiple link commands are
found in PATH, identify the first one which is a Microsoft Linker and
alter PATH in Makefile to include its directory first.