Add make lib-pkg
If a bootstrap compiler is built, make lib-pkg may be run before configure to build all the required dependencies as findlib packages as an alternative to running make lib-ext after configure.

The retrieval of the archives has much in common with the lib-ext method, and the two mechanisms are designed to live in harmony in the same Makefile. The patches directory is now ordered so that src_ext/patches/foo are the patches for foo when built with lib-ext (so as before) and src_ext/patches/foo.pkg are the patches for foo when built with lib-pkg. Additionally, src_ext/patches/foo.common are used with both mechanisms.

The URLs and MD5 checksums have been moved to src_ext/Makefile.sources and the build instructions for each package are placed in src_ext/Makefile.packages.

src_ext/Makefile has been updated to work without ../Makefile.config and Makefile cookies are used to ensure that the clone and clone-pkg targets correctly set-up the build directories.

make lib-pkg supports parallel make.