Fix OpamSystem.install for Windows
The install command doesn't make sense for Windows. OpamSystem.install
altered to use copying instead and also various checks added to guard
against package errors. In particular, .exe is automatically added to
executables (with a warning) if exec is true and warnings are displayed
if Cygwin-linked files or Unix shell-scripts are installed.