%{<...}% path expansion notation
There are many occasions in packaging of the form
%{some-directory-variable}%/more/path/components which result in an
amateur/confused-looking mixed back/forward-slash path on Windows.
String interpolation is compatibly extended such that %{<subst}% has
subst expanded with / converted to the path separator for the OS and
$var$ expanded to the value of var (with $$ being permitted to escape a
single dollar sign). Thus %{$lib$/stublibs}% is expanded to the value of
lib with "\\stublibs" on Windows.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>