Ensure versions don't conflict in packages.dev
Upstream archives are downloaded to packages.dev/$name which creates
problems during pinning when the directory is used simultaneously as a
pin directory and to hold the tarball (both sync operations are capable
of deleting the other's files).

The sanitising operations for downloading archives also mean that during
an upgrade/downgrade, one task will typically delete the other's
tarball.

Fixed by ensuring that tarballs are always placed in
packages.dev/$name.$version and packages.dev/$name is only ever used for
a pin.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>