Fix opam upgrade
The compiler will almost always be upgradable, but prevented by policy.
OPAM incorrectly displays a message that --verbose will reveal packages
which can be upgraded, but specifying this gives no detail.

Fix causes opam upgrade to display that everything is up-to-date if the
only packages which could be upgraded are part of the base.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>