Add --cc, --libc and --arch options
New options for opam init and opam switch which allow new non-system
switches to have the target C Compiler, C Runtime Library and CPU
architecture specified. These Windows-specific options allow the six
Windows ports of OCaml to be selected.

The values of these parameters (which default to the special value
system for a system switch) are available as the variables switch-cc,
switch-libc and switch-arch while the base packages are installed but
they are deleted after the switch is creation. The base packages must
therefore take responsibility for caching their values if they are
wanted subsequently (this is done by the compiler package in
opam-repository).

Signed-off-by: David Allsopp <david.allsopp@metastack.com>