Add --cc, --libc and --arch options
New options for opam init and opam switch which allow new non-system
switches to have the target C Compiler, C Runtime Library and CPU
architecture specified. These Windows-specific options allow the six
Windows ports of OCaml to be selected.

The values of these parameters are available as the variables switch-cc,
switch-libc and switch-arch. They take the special value "default" if
they have not been specified on the command line, allowing the
ocaml.system package to determine whether its installation is
appropriate (i.e. that the user has not attempted to specify the switch
configuration, making probing the system totally inappropriate).

Signed-off-by: David Allsopp <david.allsopp@metastack.com>