Fix empty values in environment file
Empty strings cause corrupt environment files, since two tabs are
written in a row which are of course lexed as a single piece of
whitespace.

Alter the format so that the @ symbol on its own means an empty string
(I'd have used something closer to $\epsilon$, but it'd never be
portable...). The single string @ is represented by \@.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>