Always install compiler package variables globally
Always copy the variables of a package flagged compiler to the switch
global variables, regardless of whether the package is being installed
to the base. This allows creating an empty switch and then adding the
packages subsequently (although it would probably better to have opam
install augmented with a command for updating the base packages and a
similar warning for installing a compiler package as a normal root as
exists for installing a normal package as a base package).

Signed-off-by: David Allsopp <david.allsopp@metastack.com>