Improved handling of overwritten files
Packages can overwrite files in shared directories (bin, sbin, etc.)
with the effect that uninstalling breaks the previous package.

This patch augments the .install mechanism with change tracking (old
versions are stored in $meta/prev). Obviously it can be circumvented
with install commands, but it allows a .install to overwrite a file on
installation but restore the original file on uninstallation.

Handling included for dealing with packages being removed in a different
order - the process should never cause files to reappear when the
underlying package is uninstalled.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>