WIP: allow variables in %{%var?true:false}%
This was part of an idea to change the ? string expansion operator so
that literal strings were enclosed in single quotes (with no embedded
expansion) and then non-quoted strings were treated as variables.

So far, so good - except that Louis and I forgot when discussing it that
you might want to use a scope on either variable, at which point the :
becomes ambiguous.

  _?scope:var:scope2:var2
  _?scope:var:var2  \ These two cases are indistinguishable in the
  _?var:scope2:var2 / general case.
  _?var:var2

A solution might be to use $ to prefix variables and keep strings as
being normal (similar to the proposed notation for %{<$lib/module}%) as
this would then permit LR parsing (Louis existing parser, as well as the
one in use here, is RL)