WIP: add printers to opam-admin.top
This was one printer intended for #install_printer. This commit is more
of a parking of an idea - it would be great if opam-admin.top (or even
the core library) had printers for all the types and installed them. It
must be possible to use the ocaml toplevel library itself to automate
the #install_printer directive?