Add switch-specific global variables to opamrc
Add a new format for switch-specific global variables along the lines of
the the "global-variables" field in /etc/opamrc. This new field
"switch-variables" is added to a configuration type "switch-defaults"
which may be embedded as a section in /etc/opamrc.

The motiviation here is to be able to specific switch-specific global
variables which can be fed to opam switch create either by the command
line or by specifying an alternate configuration file.

Signed-off-by: David Allsopp <david.allsopp@metastack.com>