WIP: Allow upgrading dependencies of the base roots
WIP NOTE: the upgrade is not explicit at the moment, and there's no
reason to be limiting this to the dependencies - it would be very useful
now to be able to switch the compiler package explicitly.

Explicit upgrade (either via opam upgrade foo, opam install foo
--upgrade or opam install foo.x.y) of dependencies of the root base
packages permitted, and messages updated accordingly.

In particular, on Windows this means that flexdll can be upgraded (and
the switch will automatically recompile itself).

Signed-off-by: David Allsopp <david.allsopp@metastack.com>