Add --text parameter to git diff
It's often advisable to add patch files to git repos as binary to
prevent CRLF mangling for files which actually require it. This would
cause git diff to display a message instead of a diff; passing --text
causes it to treat the patch file (correctly) as text and attempt to
generate a diff but it (should) still not attempt to do CRLF
normalisation.