Make hash function LF/CRLF-agnostic
Git supports (and actively uses) conversion of line endings from
LF<->CRLF. For opam, this means that a file may have had a checksum
computed for one line ending but when the file is checked out on another
system it may then have the other.

This process can obviously be ignored for files which are self-evidently
binary or which are likely to be binary (based on filename - e.g. .zip,
.tar.gz, etc.).

The implementation here is very inefficient, as it would obviously be
much better to have the two possible hashes computed over just one pass
of the file.