PR#5417: Regenerate environment file when a switch is moved
If a local switch is renamed, the environment file needs to be
recomputed. Previously, opam only patched OPAM_SWITCH_PREFIX (to ensure
the faulty environment changes could be reverted) but didn't regenerate
the file because env_hooks would be unable to use it.

Now that the environment file is written atomically and so the hooks are
not read-only, the environment is discarded if it contains an incorrect
OPAM_SWITCH_PREFIX.

The most visible effect is that the env_hook now correctly updates PATH
when a local switch is renamed. Obviously this has no effect on the
broken OCaml compiler underneath!