PR#5417: Introduce OPAM_LAST_ENV for precise env reverts
Previously, OPAM_SWITCH_PREFIX provides a best-effort for reverting the
environment variable updates made by opam. This works for most cases,
but it relies on switches having a consistent `setenv` fields and on
packages not using `setenv` fields as otherwise updates will be lost.

For example, if package `foo.1` includes `setenv: [FOO = "bar"]` then
`opam install foo.1` will cause `FOO = "bar"` to be added to
`environment`. However, `opam remove foo.1` then removes this line,
meaning that the next `opam env` will not revert `FOO=bar` from the
environment, since the `environment` file no longer contains the update.

An additional variable is now introduced, and used in `opam env` and
`opam exec`. The precise list of environment updates are stored in a
file in `/tmp/opam-last-env/` and the precise location of the file is
stored in `OPAM_LAST_ENV`. On Unix, these files should be automatically
reaped at reboot; on Windows, files older than the system uptime are
automatically pruned.

When finding the list of updates which need to be reverted, opam first
checks to see if `OPAM_LAST_ENV` points to a parseable file, falling
back to the existing `OPAM_SWITCH_PREFIX` mechanism, if required.

Hashing is used both to ensure that terminals with different active
switches do not interfere with other and also to share the environment
files between identical sessions.

With this mechanism, `opam env --revert` always succeeds and packages
are free to use `setenv`.