Pre-solve opam install --deps-only
For --deps-only, perform a top-level solve in order to make the request
precise in terms of versions. For the synthesis, this means that only
one deps-of-xxx package will ever be generated.

Some of the error messages become a little precise, at the cost of two
calls to the solver (the second of which should be more trivial than the
first).